feat: move util/uen-validation to root shared folder
