+++ src/public/modules/forms/base/directives/validate-uen.client.directive.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const { isUenValid } = require('../../../../../shared/util/uen-validation')
+const { isUenValid } = require('../../../../../../shared/util/uen-validation')
 
 angular.module('forms').directive('validateUen', validateUen)
